Made templates/game.html into a Jinja2 template
Adjusted templates/base.html to allow for custom styles and scripts on sub-templates, and converted templates/game.html to inherit structure from base and added navbar, which has a shiny new button to take you to the game from anywhere on the site. Also modified canvas dimensions to account for the navbar.